streamlit/requirements.txt
databricks-connect==16.0.0
databricks-sdk[openai]==0.46.0
databricks-sql-connector==4.0.0
pandas==2.2.3
streamlit==1.41.1
streamlit_folium==0.23.2
folium==0.18.0